Add documentation for QueueReader.
I added `QueueReader` as a section under `Software Transational
Memory`, but it is all alone.  Some other plugins actually use STM,
and maybe they should actually go in that section too.

I didn't make a safe versions since the plugin is not readable.  It
still would be nice to have, but I think exposing `stripActions` is
the appropriate solution.  That should be handled in a new PR I guess.